# docs/protocol/console-openapi-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RACP Gateway",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/devices": {
      "get": {
        "summary": "Devices",
        "operationId": "devices_api_v1_devices_get",
        "parameters": [
          {
            "name": "state",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^(ONLINE|OFFLINE|DEGRADED|CONNECTING|REVOKED)$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 2048
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListPage_DeviceView_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/operations": {
      "post": {
        "summary": "Execute",
        "operationId": "execute_api_v1_operations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OperationInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DeviceView": {
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^[A-Za-z0-9_-]{1,96}$",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "owner_id": {
            "type": "string",
            "pattern": "^[A-Za-z0-9_-]{1,96}$",
            "title": "Owner Id"
          },
          "revoked": {
            "type": "integer",
            "title": "Revoked"
          },
          "epoch": {
            "type": "integer",
            "title": "Epoch"
          },
          "info": {
            "additionalProperties": true,
            "type": "object",
            "title": "Info"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "name",
          "owner_id",
          "revoked",
          "epoch",
          "info"
        ],
        "title": "DeviceView"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ListPage_DeviceView_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DeviceView"
            },
            "type": "array",
            "maxItems": 500,
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "consistency": {
            "type": "string",
            "const": "best_effort",
            "title": "Consistency",
            "default": "best_effort"
          },
          "sort": {
            "type": "string",
            "const": "rowid_desc",
            "title": "Sort",
            "default": "rowid_desc"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "ListPage[DeviceView]"
      },
...
      "OperationInput": {
        "properties": {
          "device_id": {
            "type": "string",
            "pattern": "^[A-Za-z0-9_-]{1,96}$",
            "title": "Device Id"
          },
          "workspace_id": {
            "type": "string",
            "pattern": "^[A-Za-z][A-Za-z0-9_-]{0,31}$",
            "title": "Workspace Id",
            "default": "default"
          },
          "operation": {
            "type": "string",
            "maxLength": 96,
            "pattern": "^[a-z]+\\.[a-z_]+$",
            "title": "Operation"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          },
          "timeout_ms": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 86400000.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Timeout Ms"
          },
          "execution_mode": {
            "type": "string",
            "enum": [
              "sync",
              "job"
            ],
            "title": "Execution Mode",
            "default": "sync"
          },
          "idempotency_key": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128,
                "minLength": 1,
                "pattern": "^[!-~]+$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          },
          "execution_profile_id": {
            "type": "string",
            "enum": [
              "read_only",
              "standard",
              "trusted_personal"
            ],
            "title": "Execution Profile Id",
            "default": "read_only"
          },
          "approval_id": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[A-Za-z0-9_-]{1,96}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approval Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "device_id",
          "operation",
          "payload"
        ],
        "title": "OperationInput"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}